// propertyApp/src/app/protected/interfaces/iente.ts
export interface IEnte {
    id: string;
    accountid: string;
    name: string;
    gender: string;
    nationality: string;
    doctype: string;
    docnumber: string;
    dob: string;
    dni: string;
    fiscalname: string;
    fulladdress: string;
    email: string;
    phone: string;
    image: string;
    status: string;
    created_by?: string;
    created_at?: string;
    modified_by?: string;
    modified_at?: string;
}