// propertyApp/src/app/protected/interfaces/isale.ts
export interface ISale {
  id: string;
  accountid: string;
  enteid: string;
  itemid: string;
  currency: string;
  parity: string;
  taxname: string;
  taxamount: string;
  subtotal: string;
  total: string;
  paid: string;
  day: string;
  interestrate: string;
  daysoftolerance: string;
  guaranteedeposit: string;
  type: string;
  startdate: string;
  enddate: string;
  duration: string;
  comments: string;
  status: string;
  created_by: string;
  created_at: string;
  modified_by: string;
  modified_at: string;
}